Added conditional HTML tag, HTML5 Shiv, and Respond.js scripts
The comment characters around the HTML5 Shiv, Respond.js and the wrapping HTML tag are to be retained whenever you add these to the document. This ensures that the code is only rendered when the conditions stated are met. For instance, in the code below, the part that reads <!--[if lt IE 9]> means that if the browser detected is less than Internet Explorer 9, then the scripts will be loaded. If the IE version is greater than IE9, the scripts will not render.

<!--[if lt IE 9]>
      <script src="https://oss.maxcdn.com/html5shiv/3.7.3/html5shiv.min.js"></script>
      <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
    <![endif]-->